# docs/sms-api-doc.yaml
openapi: 3.0.1
info:
  title: School Management System API
  version: "1.3.3"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /schedules/{scheduleId}/CANCEL:
    put:
      tags:
      - Schedules
      summary: Cancel schedule
      description: "Cancel booked schedule and all it's occurrences. Required authority\
        \ level: Administrator"
      operationId: cancelSchedule
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              maxLength: 200
              minLength: 3
              type: string
        required: true
      responses:
        "404":
          description: Schedule not found
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: Canceled successfully
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Full authentication is required to access this resource
        "406":
          description: Cancellation not possible because of schedule's status
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
...
  /schedules/version/{scheduleVersionId}:
    get:
      tags:
      - Schedules
      summary: Get schedule version by ID
      description: "Retrieves existing schedule version for given ID. Required authority\
        \ level: Administrator, Teacher, Student"
      operationId: getScheduleVersionsForId
      parameters:
      - name: scheduleVersionId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Full authentication is required to access this resource
        "200":
          description: Returns existing schedule version
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleVersionDetailedResponse'
        "406":
          description: Schedule version already marked as removed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleVersionDetailedResponse'
        "403":
          description: User lacks permission to access this resource
        "404":
          description: Schedule version not found for given ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleVersionDetailedResponse'
      security:
      - bearerAuth: []
    post:
      tags:
      - Schedules
      summary: Create new schedule
      description: "Adds new schedule representing recurrent lesson unit applicable\
        \ in a given time period. Required authority level: Administrator"
      operationId: addSchedule
      parameters:
      - name: scheduleVersionId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddScheduleRequest'
        required: true
      responses:
        "404":
          description: Schedule version not found by ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
        "201":
          description: Created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "406":
          description: Teacher not available in a given time period
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
    delete:
      tags:
      - Schedules
      summary: Delete schedule version
      description: "Remove existing schedule version with given ID, if not already\
        \ marked as removed. Required authority level: Administrator"
      operationId: deleteScheduleVersion
      parameters:
      - name: scheduleVersionId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Removed successfully
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Full authentication is required to access this resource
        "406":
          description: Schedule version already marked as removed
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: User lacks permission to access this resource
        "404":
          description: Not found by ID
          content:
            '*/*':
              schema:
                type: string
      security:
      - bearerAuth: []
    patch:
      tags:
      - Schedules
      summary: Update schedule version
      description: "Modify existing schedule version. Required authority level: Administrator"
      operationId: patchScheduleVersion
      parameters:
      - name: scheduleVersionId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchScheduleVersionRequest'
        required: true
      responses:
        "401":
          description: Full authentication is required to access this resource
        "404":
          description: Schedule version not found by ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleVersionDetailedResponse'
        "200":
          description: Modified successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleVersionDetailedResponse'
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
  /schedules/version/{scheduleVersionId}/clone:
    post:
      tags:
      - Schedules
      summary: Clone schedule version
      description: "Creates new schedule version based on date in cloned version with\
        \ all schedules, that is set to unactive by default. Required authority level:\
        \ Administrator"
      operationId: cloneScheduleVersion
      parameters:
      - name: scheduleVersionId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Schedule version to clone not found by ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleVersionDetailedResponse'
        "201":
          description: Successfully created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleVersionDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
        "406":
          description: Schedule version to clone is already marked as removed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleVersionDetailedResponse'
      security:
      - bearerAuth: []
...
  /schedules/{scheduleId}:
    delete:
      tags:
      - Schedules
      summary: Delete schedule
      description: "Remove existing schedule if has no ongoing occurrences, and cancels\
        \ booked occurrences. Required authority level: Administrator"
      operationId: deleteSchedule
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "406":
          description: Active occurrences ongoing
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: Removed successfully
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
        "404":
          description: Not found by ID
          content:
            '*/*':
              schema:
                type: string
      security:
      - bearerAuth: []
    patch:
      tags:
      - Schedules
      summary: Update schedule
      description: "Modify existing schedule, if time period is applicable and already\
        \ booked. Required authority level: Administrator"
      operationId: patchSchedule
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchScheduleRequest'
        required: true
      responses:
        "406":
          description: Modification canceled because of actual schedule status
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
        "200":
          description: Modification successful
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "404":
          description: Not found by ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
...
  /schedules/{scheduleVersionId}/{scheduleId}:
    get:
      tags:
      - Schedules
      summary: Find schedule by ID
      description: "Find existing schedule with ID. Required authority level: Administrator,\
        \ Teacher, Student"
      operationId: getSchedule
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: scheduleVersionId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Returns existing object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
        "404":
          description: Not found by ID
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleDetailedResponse'
      security:
      - bearerAuth: []
  /schedules/version/{scheduleVersionId}/day/{dayOfWeek}:
    get:
      tags:
      - Schedules
      summary: Get list of schedules for day of week
      description: "Find all schedules booked on particular day of week for schedule\
        \ version. Required authority level: Administrator, Teacher, Student"
      operationId: getScheduleList
      parameters:
      - name: scheduleVersionId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: dayOfWeek
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "401":
          description: Full authentication is required to access this resource
        "200":
          description: Returns existing objects (empty result as well)
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleSummaryResponse'
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
  /schedules/version/yearbook/{yearbookId}:
    get:
      tags:
      - Schedules
      summary: Get all schedule versions for yearbook
      description: "Retrieves all schedule versions for Yearbook that are not marked\
        \ as removed. Required authority level: Administrator"
      operationId: getScheduleVersionsForYearbook
      parameters:
      - name: yearbookId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Returns list of objects (empty result as well)
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleVersionSummaryResponse'
        "401":
          description: Full authentication is required to access this resource
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
  /schedules/version/yearbook/{yearbookId}/active:
    get:
      tags:
      - Schedules
      summary: Get active schedule version for Yearbook
      description: "Retrieves active schedule version for Yearbook. Required authority\
        \ level: Administrator, Teacher, Student"
      operationId: getActiveScheduleVersionsForYearbook
      parameters:
      - name: yearbookId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "406":
          description: No active schedule version for given yearbook
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleVersionDetailedResponse'
        "401":
          description: Full authentication is required to access this resource
        "200":
          description: Returns existing active schedule version for given yearbook
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScheduleVersionDetailedResponse'
        "403":
          description: User lacks permission to access this resource
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    YearbookSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        symbol:
          type: string
        headTeacher:
          $ref: '#/components/schemas/TeacherSummaryResponse'
        startingYear:
          type: string
          format: date-time
        graduationYear:
          type: string
          format: date-time
...
    ScheduleDetailedResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        subject:
          $ref: '#/components/schemas/SubjectSummaryResponse'
        teacher:
          $ref: '#/components/schemas/TeacherSummaryResponse'
        classroom:
          $ref: '#/components/schemas/ClassroomSummaryResponse'
        dayOfWeek:
          type: string
          enum:
          - MONDAY
          - TUESDAY
          - WEDNESDAY
          - THURSDAY
          - FRIDAY
          - SATURDAY
          - SUNDAY
        startTime:
          $ref: '#/components/schemas/LocalTime'
        endTime:
          $ref: '#/components/schemas/LocalTime'
        recurrenceType:
          type: string
          enum:
          - NONE
          - WEEKLY
          - BIWEEKLY
          - MONTHLY
        scheduleVersion:
          $ref: '#/components/schemas/ScheduleVersionSummaryResponse'
        status:
          type: string
          enum:
          - SCHEDULED
          - CANCELLED
          - RESCHEDULED
          - DELETED
          - UPDATED
        effectiveDate:
          type: string
          format: date
        expirationDate:
          type: string
          format: date
    ScheduleVersionDetailedResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        createDate:
          type: string
          format: date-time
        yearbook:
          $ref: '#/components/schemas/YearbookSummaryResponse'
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleDetailedResponse'
        active:
          type: boolean
    ScheduleVersionSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        active:
          type: boolean
    AddScheduleRequest:
      required:
      - endTime
      - recurrenceType
      - startTime
      type: object
      properties:
        subjectId:
          maximum: 2147483647
          minimum: 1
          type: integer
          format: int32
        teacherId:
          maximum: 2147483647
          minimum: 1
          type: integer
          format: int32
        classroomId:
          maximum: 2147483647
          minimum: 1
          type: integer
          format: int32
        dayOfWeek:
          maximum: 7
          minimum: 1
          type: integer
          format: int32
        startTime:
          $ref: '#/components/schemas/LocalTime'
        endTime:
          $ref: '#/components/schemas/LocalTime'
        recurrenceType:
          type: string
          enum:
          - NONE
          - WEEKLY
          - BIWEEKLY
          - MONTHLY
        effectiveDate:
          type: string
          format: date
        expirationDate:
          type: string
          format: date
        validExpirationDate:
          type: boolean
        expirationDateIfRecurrenceNone:
          type: boolean
        effectiveDateNotNull:
          type: boolean
        effectiveDateBeforeExpirationDate:
          type: boolean
        validTimeRange:
          type: boolean
...
    PatchScheduleVersionRequest:
      type: object
      properties:
        name:
          type: string
        yearbookId:
          type: integer
          format: int32
        active:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT